backend/requirements.txt
fastapi[standard]
uvicorn[standard]
redis[hiredis]

python-dotenv
pydantic-settings

pytest
pytest-asyncio
httpx